driver-management-service/requirements.txt
Flask
redis
kafka-python